Remove database module
