+++ static/app/views/starfish/modules/DBModule.tsx
@@ -1,0 +1,32 @@
+import * as Layout from 'sentry/components/layouts/thirds';
+import PageFiltersContainer from 'sentry/components/organizations/pageFilters/container';
+import {t} from 'sentry/locale';
+import {
+  PageErrorAlert,
+  PageErrorProvider,
+} from 'sentry/utils/performance/contexts/pageError';
+import {ModuleName} from 'sentry/views/starfish/types';
+import SpansView from 'sentry/views/starfish/views/spans/spansView';
+
+export default function DBModule() {
+  return (
+    <Layout.Page>
+      <PageErrorProvider>
+        <Layout.Header>
+          <Layout.HeaderContent>
+            <Layout.Title>{t('DB')}</Layout.Title>
+          </Layout.HeaderContent>
+        </Layout.Header>
+
+        <Layout.Body>
+          <Layout.Main fullWidth>
+            <PageErrorAlert />
+            <PageFiltersContainer>
+              <SpansView moduleName={ModuleName.DB} />
+            </PageFiltersContainer>
+          </Layout.Main>
+        </Layout.Body>
+      </PageErrorProvider>
+    </Layout.Page>
+  );
+}
